% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shard_calculator}
\alias{shard_calculator}
\title{Calculate the Number of Shards and Simultaneous Chains}
\usage{
shard_calculator(num_respondents, num_responses, num_chains)
}
\arguments{
\item{num_respondents}{An integer specifying the number of respondents.}

\item{num_responses}{An integer specifying the number of responses.}

\item{num_chains}{An integer specifying the number of chains that need to be
run.}
}
\value{
`ret` A list containing the number of shards to use within each chain
and the number of chains to run in parallel.
}
\description{
Calculating the number of shards and simultaneous chains.
}
\examples{
shard_calculator(num_respondents = 1000, num_responses = 5000, num_chains = 4)
}
