% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-fng-tdcm-stan.R
\name{create_fng_stan_tdcm}
\alias{create_fng_stan_tdcm}
\title{Creating Fungible TDCM Stan Code}
\usage{
create_fng_stan_tdcm(q_matrix)
}
\arguments{
\item{q_matrix}{A tibble containing the assessment Q-matrix.}
}
\value{
`stan_code` A list containing the text for the Stan code blocks.
}
\description{
Automating the creation of fungible Stan code for a TDCM.
}
\examples{
qmatrix = tibble::tibble(att_1 = c(1, 0, 1, 0, 1, 1), att_2 = c(0, 1, 0, 1, 1, 1))
create_fng_stan_tdcm(q_matrix = qmatrix)
}
